support for arrays on function post
Parameters added:
sanitize: "filename", "search", "html" or "sql". All options are, in one
or other time, filtered by XSS_filter.
check_XID: "no" as default. In case of "yes", check for the post of a
valid XID and returns empty em case of false.
$glue: implode the array using this parameter as glue. For traditional
use as single variable.
split_by: in case of using as variable pair and post of string, explodes
the variable.

Variable: surg:post:value, in case of using as variable pair.

Conflicts:
	pi.surgeree.php